+++ evm/src/main/java/org/hyperledger/besu/evm/gascalculator/BerlinGasCalculator.java
@@ -164,17 +164,15 @@ public class BerlinGasCalculator extends IstanbulGasCalculator {
   @Override
   // As per https://eips.ethereum.org/EIPS/eip-2200
   public long calculateStorageCost(
-      final Account account,
-      final UInt256 key,
       final UInt256 newValue,
       final Supplier<UInt256> currentValue,
       final Supplier<UInt256> originalValue) {
 
-    UInt256 localCurrentValue = currentValue.get();
+    final UInt256 localCurrentValue = currentValue.get();
     if (localCurrentValue.equals(newValue)) {
       return SLOAD_GAS;
     } else {
-      UInt256 localOriginalValue = originalValue.get();
+      final UInt256 localOriginalValue = originalValue.get();
       if (localOriginalValue.equals(localCurrentValue)) {
         return localOriginalValue.isZero() ? SSTORE_SET_GAS : SSTORE_RESET_GAS;
       } else {
@@ -187,17 +185,15 @@ public class BerlinGasCalculator extends IstanbulGasCalculator {
   @Override
   // As per https://eips.ethereum.org/EIPS/eip-2200
   public long calculateStorageRefundAmount(
-      final Account account,
-      final UInt256 key,
       final UInt256 newValue,
       final Supplier<UInt256> currentValue,
       final Supplier<UInt256> originalValue) {
 
-    UInt256 localCurrentValue = currentValue.get();
+    final UInt256 localCurrentValue = currentValue.get();
     if (localCurrentValue.equals(newValue)) {
       return 0L;
     } else {
-      UInt256 localOriginalValue = originalValue.get();
+      final UInt256 localOriginalValue = originalValue.get();
       if (localOriginalValue.equals(localCurrentValue)) {
         if (localOriginalValue.isZero()) {
           return 0L;